+++ src/libs/io.h
@@ -10,12 +10,12 @@ namespace image_utils {
 
     void write_image(image_RGB &rgb_data, const std::string &out_filename);
 
-    void image_sanity_check(const matrix<long double> &doubles,
+    void image_sanity_check(const matrix<double> &doubles,
                             bool print_minmax = false);
 
-    void color_write_image(matrix<long double> &grid, colormap *cmap,
+    void color_write_image(matrix<double> &grid, colormap *cmap,
                            const std::string &output_filename);
 
-    void scale_grid(matrix<long double> &doubles);
+    void scale_grid(matrix<double> &doubles);
 
 }
