+++ src/tests/test_generators.cpp
@@ -8,7 +8,7 @@ int main(int argc, char const *argv[]) {
     using namespace image_utils;
     {
         wave *w = new wave_square();
-        long double test_values[][2] = {
+        double test_values[][2] = {
                 {0,0},
                 {0.4, 0},
                 {0.5,1},
@@ -16,13 +16,13 @@ int main(int argc, char const *argv[]) {
                 {0.999,1},
                 {1,0}
         };
-        for (long double *d : test_values) {
+        for (double *d : test_values) {
             test((*w)(d[0]) == d[1], "square wave");
         }
     }
     {
         wave *w = new wave_triangle();
-        long double test_values[][2] = {
+        double test_values[][2] = {
                 {0,0},
                 {0.375,2*0.375},
                 {0.25,0.5},
@@ -30,7 +30,7 @@ int main(int argc, char const *argv[]) {
                 {0.75,0.5},
                 {1,0}
         };
-        for (long double *d : test_values) {
+        for (double *d : test_values) {
             test((*w)(d[0]) == d[1], "triangle wave");
             if ((*w)(d[0]) != d[1]) {
                 debug_log((*w)(d[0]));
@@ -40,13 +40,13 @@ int main(int argc, char const *argv[]) {
     }
     {
         wave *w = new wave_sawtooth();
-        long double test_values[][2] = {
+        double test_values[][2] = {
                 {0,0},
                 {0.1,0.1},
                 {0.9,0.9},
                 {4.1,0.1}
         };
-        for (long double *d : test_values) {
+        for (double *d : test_values) {
             test(std::fabs((*w)(d[0]) - d[1]) < 0.0001, "sawtooth wave");
             if (std::fabs((*w)(d[0]) - d[1]) > 0.0001) {
                 debug_log((*w)(d[0]));
